refactor: split normalization from validation
